Move other incorrectly placed maintained app method
