+++ ee/server/service/maintained_apps.go
@@ -102,6 +102,21 @@ func (svc *Service) AddFleetMaintainedApp(ctx context.Context, teamID *uint, app
 	return nil
 }
 
+func (svc *Service) ListFleetMaintainedApps(ctx context.Context, teamID uint, opts fleet.ListOptions) ([]fleet.MaintainedApp, *fleet.PaginationMetadata, error) {
+	if err := svc.authz.Authorize(ctx, &fleet.AuthzSoftwareInventory{
+		TeamID: &teamID,
+	}, fleet.ActionRead); err != nil {
+		return nil, nil, err
+	}
+
+	avail, meta, err := svc.ds.ListAvailableFleetMaintainedApps(ctx, teamID, opts)
+	if err != nil {
+		return nil, nil, ctxerr.Wrap(ctx, err, "listing available fleet managed apps")
+	}
+
+	return avail, meta, nil
+}
+
 func (svc *Service) GetFleetMaintainedApp(ctx context.Context, appID uint) (*fleet.MaintainedApp, error) {
 	if err := svc.authz.Authorize(ctx, &fleet.AuthzSoftwareInventory{
 		TeamID: nil,
