+++ ee/server/service/software.go
@@ -3,7 +3,6 @@ package service
 import (
 	"context"
 
-	"github.com/fleetdm/fleet/v4/server/contexts/ctxerr"
 	"github.com/fleetdm/fleet/v4/server/fleet"
 )
 
@@ -16,19 +15,4 @@ func (svc *Service) ListSoftware(ctx context.Context, opts fleet.SoftwareListOpt
 func (svc *Service) SoftwareByID(ctx context.Context, id uint, teamID *uint, _ bool) (*fleet.Software, error) {
 	// reuse SoftwareByID, but include cve scores in premium version
 	return svc.Service.SoftwareByID(ctx, id, teamID, true)
-}
-
-func (svc *Service) ListFleetMaintainedApps(ctx context.Context, teamID uint, opts fleet.ListOptions) ([]fleet.MaintainedApp, *fleet.PaginationMetadata, error) {
-	if err := svc.authz.Authorize(ctx, &fleet.AuthzSoftwareInventory{
-		TeamID: &teamID,
-	}, fleet.ActionRead); err != nil {
-		return nil, nil, err
-	}
-
-	avail, meta, err := svc.ds.ListAvailableFleetMaintainedApps(ctx, teamID, opts)
-	if err != nil {
-		return nil, nil, ctxerr.Wrap(ctx, err, "listing available fleet managed apps")
-	}
-
-	return avail, meta, nil
 }
